Fix a missing edge case in day 2 part 2 solution
If the first and last number of the entry were the same and removing the
last number would fix the entry and make it safe, the previous logic
wouldn't check for removal of the last number in that line of entry.
This quick fix checks for removing the last number as long as there is a
single problem somewhere in the entry, so it's not efficient, but it
works!